@@ -9,11 +9,11 @@
     "axios": "^0.27.2",
     "bootstrap": "^5.1.3",
     "json-server": "^1.0.0-alpha.23",
-    "react": "^17.0.2",
+    "react": "^19.1.0",
     "react-bootstrap": "^2.1.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^19.1.0",
     "react-router-dom": "^6.2.2",
-    "react-scripts": "4.0.3",
+    "react-scripts": "^5.0.1",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
