@@ -1,15 +1,6 @@
-import React from 'react';
-import reactDom from 'react-dom';
-import ReactDOM from 'react-dom';
-import reportWebVitals from './reportWebVitals';
+
 import App from './App';
-
-ReactDOM.render(
-  <App />, 
-  document.getElementById('root')
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+import { createRoot } from 'react-dom/client';
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(<App />);
